added class Component
