+++ src/component.js
@@ -1,0 +1,38 @@
+import {createElement} from './utils.js';
+
+
+class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate Component, only concrete one.`);
+    }
+
+    this._element = null;
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  get element() {
+    return this._element;
+  }
+
+  render() {
+    this._element = createElement(this.template)[0];
+    this.bind();
+    return this._element;
+  }
+
+  unrender() {
+    this.unbind();
+    this._element.remove();
+    this._element = null;
+  }
+
+  bind() {}
+
+  unbind() {}
+}
+
+export {Component};
