+++ src/trip-point.js
@@ -1,8 +1,10 @@
-import {createElement, isFunction} from './utils.js';
+import {isFunction} from './utils.js';
 import {typeTripPoint} from './data.js';
+import {Component} from './component.js';
 
-class TripPoint {
+class TripPoint extends Component {
   constructor(data) {
+    super();
     this._type = data.type;
     this._destination = data.destination;
     this._day = data.day;
@@ -36,22 +38,6 @@ class TripPoint {
         ${this._offers.map((offer) => `<li><button class="trip-point__offer">${offer.name} +&euro;&nbsp;${offer.price}</button></li>`).join(``)}
       </ul>
     </article>`.trim();
-  }
-
-  get element() {
-    return this._element;
-  }
-
-  render() {
-    this._element = createElement(this.template)[0];
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element.remove();
-    this._element = null;
   }
 
   set onEdit(fn) {
